refactor(mapper): rename package handler to exceptionHandler
